refactor school listing data flow and optimize Supabase access
Split school listing fetch logic into a dedicated data hook and moved row rendering into a separate renderer component
Switched school listing auth lookup to supabase.auth.getUser()
Simplified school listing query to fetch full school_metrics rows with direct access filtering
Added is_deleted to school_metrics in generated database types
Centralized program type and program model validation using shared constants
Kept school-list behavior intact and verified with the school-list test suite
